handle deprecations in separate file
