+++ packages/node-resolve/src/index.js
@@ -9,6 +9,7 @@ import { isDirCached, isFileCached, readCachedFile } from './cache';
 import { exists, readFile, realpath } from './fs';
 import { resolveImportSpecifiers } from './resolveImportSpecifiers';
 import { getMainFields, getPackageName, normalizeInput } from './util';
+import handleDeprecatedOptions from './deprecated-options';
 
 const builtins = new Set(builtinList);
 const ES6_BROWSER_EMPTY = '\0node-resolve:empty.js';
@@ -38,47 +39,7 @@ const defaults = {
 export const DEFAULTS = deepFreeze(deepMerge({}, defaults));
 
 export function nodeResolve(opts = {}) {
-  const warnings = [];
-
-  if (opts.customResolveOptions) {
-    const { customResolveOptions } = opts;
-    if (customResolveOptions.moduleDirectory) {
-      opts.moduleDirectories = Array.isArray(customResolveOptions.moduleDirectory)
-        ? customResolveOptions.moduleDirectory
-        : [customResolveOptions.moduleDirectory];
-
-      warnings.push(
-        'node-resolve: The `customResolveOptions.moduleDirectory` option has been deprecated. Use `moduleDirectories`, which must be an array.'
-      );
-    }
-
-    if (customResolveOptions.preserveSymlinks) {
-      throw new Error(
-        'node-resolve: `customResolveOptions.preserveSymlinks` is no longer an option. We now always use the rollup `preserveSymlinks` option.'
-      );
-    }
-
-    [
-      'basedir',
-      'package',
-      'extensions',
-      'includeCoreModules',
-      'readFile',
-      'isFile',
-      'isDirectory',
-      'realpath',
-      'packageFilter',
-      'pathFilter',
-      'paths',
-      'packageIterator'
-    ].forEach((resolveOption) => {
-      if (customResolveOptions[resolveOption]) {
-        throw new Error(
-          `node-resolve: \`customResolveOptions.${resolveOption}\` is no longer an option. If you need this, please open an issue.`
-        );
-      }
-    });
-  }
+  const { warnings } = handleDeprecatedOptions(opts);
 
   const options = { ...defaults, ...opts };
   const { extensions, jail, moduleDirectories } = options;
